fix: Eliminate "leaking this from constructor" warning.
Create a MenuListener class that contains an actionPerformed() method
that will respond to selections of items on the Background Color and 
Foreground Color methods. Swing class no longer listens for these events. 
The constructor for the Swing class no longer pays a reference to the 
instance it is constructing to the addListener() method of a JMenuItem 
item.