refactor: Eliminate duplicated and extraneous code in Prism class.
The Prism class no longer divides the polygons the define the top and
bottom of the prism into triangles. The Polygon3D now has an instance
variable whose value signals whether vertices (Vectors) are listed in 
clockwise or counterclockwise order. The method that computes a 
Polygon3D's normal Vector uses the order of the listing of Vectors to 
determine the orientation of the normal Vector.